Fix two bugs in PR properties (#4346)
* Save the refreshed PR properties, not the lookup ones

I mixed up the properties that we use to look up the PR and those that
we actually fetch from the provider. Let's save those that we use to
look up the PR to ensure that we have all the data available.

In this particular case, we were missing the SHA.

* GitHub's PR number is not int64, but our protobuf is

Our protobuf defines the PR number as int64, but I didn't notice that
the GH API returns the PR number as plain int. It's easiest to upcast
the GH value to int64.

+++ internal/providers/github/properties/pull_request.go
@@ -142,8 +142,9 @@ func getPrWrapper(
 		properties.PropertyUpstreamID: prReply.GetID(),
 		properties.PropertyName:       fmt.Sprintf("%s/%s/%d", owner, name, id),
 		// github-specific
-		PullPropertyURL:       prReply.GetHTMLURL(),
-		PullPropertyNumber:    prReply.GetNumber(),
+		PullPropertyURL: prReply.GetHTMLURL(),
+		// our proto representation uses int64 for the number but GH uses int
+		PullPropertyNumber:    int64(prReply.GetNumber()),
 		PullPropertySha:       prReply.GetHead().GetSHA(),
 		PullPropertyRepoOwner: owner,
 		PullPropertyRepoName:  name,
